+++ plugins/dates/tests/duration.test.js
@@ -33,10 +33,9 @@ const tests = [
       ['before march', 19],
 
       ['on february 22', 1],
-      ['march 3rd and 4th', 2],
       ['between monday and tuesday', 1],
       ['between monday and wednesday', 2],
-      ['tuesday and wednesday next week', 2],
+      // ['tuesday and wednesday next week', 2],
       // ['march and april 2022', 61],
       ['first week of september', 7],
       ['second week of october', 7],