+++ plugins/dates/changelog.md
@@ -1,6 +1,10 @@
 <!-- #### [Unreleased]
-
+- **[breaking]** - return array in .json().dates
 -->
+
+### 2.2.0 [April 2021]
+
+- **[new]** - start parsing or/and date combos
 
 ### 2.1.0 [April 2021]